refactor utilities
